+++ client/bindings/nodejs/types/bridge/client.ts
@@ -185,6 +185,10 @@ export interface __GetTipsIntervalMessage__ {
     name: 'getTipsInterval';
 }
 
+export interface __GetProtocolParametersMessage__ {
+    name: 'getProtocolParameters';
+}
+
 export interface __GetLocalPowMessage__ {
     name: 'getLocalPow';
 }